Fix crash when there is no files to delete

+++ app/src/main/java/com/sslabs/whatsappcleaner/CleanerReceiver.java
@@ -43,9 +43,11 @@ public class CleanerReceiver extends BroadcastReceiver {
                                 pathname.getName());
                     }
                 });
-                for (File toDelete : oldBackupDatabases) {
-                    Log.d(MainActivity.TAG, "File to delete: " + toDelete.getName());
-                    toDelete.delete();
+                if (oldBackupDatabases != null && oldBackupDatabases.length > 0) {
+                    for (File toDelete : oldBackupDatabases) {
+                        Log.d(MainActivity.TAG, "File to delete: " + toDelete.getName());
+                        toDelete.delete();
+                    }
                 }
             }
             return null;
